Moving skycoord_equal to asdf.tags.helpers
