+++ astropy/io/misc/asdf/tags/coordinates/skycoord.py
@@ -4,7 +4,7 @@
 from asdf.yamlutil import custom_tree_to_tagged_tree
 
 from astropy.coordinates import SkyCoord
-from astropy.coordinates.sky_coordinate import _skycoord_equal as skycoord_equal
+from astropy.io.misc.asdf.tags.helpers import skycoord_equal
 
 from ...types import AstropyType
